HIIT-014 rename: EntitySearchParam 패키지 이동
